+++ Luna/luna_math/trig.cpp
@@ -7,22 +7,22 @@ namespace luna
 {
 	namespace math
 	{
-		double DegreeToRadian(double angleInDegrees)
+		float DegreeToRadian(float angleInDegrees)
 		{
 			return  angleInDegrees * M_PI / 180;
 		}
 
-		double RadianToDegree(double angleInRadians)
+		float RadianToDegree(float angleInRadians)
 		{
 			return  angleInRadians * 180 / M_PI;
 		}
 
-		double ArcLengthRadians(double radius, double angleInRadians)
+		float ArcLengthRadians(float radius, float angleInRadians)
 		{
 			return angleInRadians * radius;
 		}
 
-		double ArcLengthDegrees(double radius, double angleInDegrees)
+		float ArcLengthDegrees(float radius, float angleInDegrees)
 		{
 			return (angleInDegrees / 360) * 2 * M_PI * radius;
 		}
